requirements.txt
Django==2.2.7
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3
sqlparse==0.3.0
